+++ example/main.tf
@@ -7,7 +7,6 @@ provider "tls" {
 }
 
 provider "google" {
-  version = "~> 2.20.0"
   project = var.project
   region  = var.zone
 }
@@ -18,23 +17,28 @@ provider "google-beta" {
 }
 
 module "elasticsearch_prod" {
-  source            = "./.."
-  project           = var.project
-  region            = var.region
-  zone              = var.zone
-  instance_name     = "elasticsearch-prod"
-  cluster_name      = "elasticsearch"
-  cluster_ipv4_cidr = module.gke.cluster_ipv4_cidr
-  node_count        = "3"
-  heap_size         = "1500m"
-  raw_image_source  = "https://storage.googleapis.com/ackee-images/ackee-elasticsearch-7-disk-79.tar.gz"
-  data_disk_size    = "10"
-  k8s_enable        = true
-  namespace         = var.namespace
+  source  = "./.."
+  project = var.project
+  region  = var.region
+  zone    = var.zone
+
+  instance_name    = "elasticsearch-prod"
+  cluster_name     = "elasticsearch"
+  node_count       = 2
+  heap_size        = "1500m"
+  raw_image_source = "https://storage.googleapis.com/ackee-images/ackee-elasticsearch-7-disk-79.tar.gz"
+  data_disk_size   = "10"
+
+  namespace = var.namespace
+
+  cluster_ca_certificate = module.gke.cluster_ca_certificate
+  cluster_user           = module.gke.cluster_username
+  cluster_password       = module.gke.cluster_password
+  cluster_endpoint       = module.gke.endpoint
 }
 
 module "gke" {
-  source            = "git::ssh://git@gitlab.ack.ee/Infra/terraform-gke-vpc.git?ref=v6.2.0"
+  source            = "git::ssh://git@gitlab.ack.ee/Infra/terraform-gke-vpc.git?ref=v6.4.0"
   namespace         = var.namespace
   project           = var.project
   location          = var.zone
@@ -61,4 +65,12 @@ variable "region" {
 
 variable "zone" {
   default = "europe-west3-c"
+}
+
+output "ip_address" {
+  value = module.elasticsearch_prod.ip_address
+}
+
+output "ilb_dns" {
+  value = module.elasticsearch_prod.ilb_dns
 }